added 'days completed' endpoint

+++ API/domain/time_entry.go
@@ -8,7 +8,7 @@ type TimeEntryModel struct {
 	BaseModel
 	StartTimeUtc   time.Time `gorm:"column:StartTimeUtc"`
 	EndTimeUtc     time.Time `gorm:"column:EndTimeUtc"`
-	DailyWorkHours float32   `gorm:"column:DailyWorkHours"`
+	DailyWorkHours float64   `gorm:"column:DailyWorkHours"`
 	PauseSeconds   int       `gorm:"column:PauseSeconds"`
 	Note           string    `gorm:"column:Note"`
 	IsDeleted      bool      `gorm:"column:IsDeleted"`
